requirements.txt
streamlit
streamlit-webrtc
tensorflow-cpu
opencv-python
mediapipe
numpy
pandas
scikit-learn
pickle-mixin